Rename test helper file so that it is not treated as a test file
