+++ client/shared/src/components/ranking/ZoektRanking.test.tsx
@@ -1,4 +1,4 @@
-import { testDataRealMatches } from './PerFileResultRanking.test'
+import { testDataRealMatches } from './PerFileResultRankingTestHelpers'
 import { ZoektRanking } from './ZoektRanking'
 
 describe('ZoektRanking', () => {
